🚚 move : 에러 응답 파일 이동
