+++ app/src/main/java/com/yomo/yakgwa/ui/myPage/MyPageViewModel.kt
@@ -4,17 +4,17 @@ import android.net.Uri
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.yomo.data.ErrorResponse
+import com.skydoves.sandwich.onSuccess
+import com.skydoves.sandwich.retrofit.serialization.onErrorDeserialize
+import com.skydoves.sandwich.suspendOnSuccess
 import com.yomo.data.datasource.local.YakGwaLocalDataSource
+import com.yomo.data.utils.ErrorResponse
 import com.yomo.domain.model.response.UserInfoResponseEntity
 import com.yomo.domain.usecase.GetUserInformationUseCase
 import com.yomo.domain.usecase.PatchUpdateUserImageUseCase
 import com.yomo.domain.usecase.PostUserLogoutUseCase
 import com.yomo.domain.usecase.PostUserSignoutUseCase
 import com.yomo.yakgwa.util.UiState
-import com.skydoves.sandwich.onSuccess
-import com.skydoves.sandwich.retrofit.serialization.onErrorDeserialize
-import com.skydoves.sandwich.suspendOnSuccess
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
