+++ app/src/main/java/com/yomo/yakgwa/ui/promiseHistory/PromiseHistoryViewModel.kt
@@ -2,12 +2,12 @@ package com.yomo.yakgwa.ui.promiseHistory
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.yomo.data.ErrorResponse
+import com.skydoves.sandwich.onSuccess
+import com.skydoves.sandwich.retrofit.serialization.onErrorDeserialize
+import com.yomo.data.utils.ErrorResponse
 import com.yomo.domain.model.response.PromiseHistoryResponseEntity
 import com.yomo.domain.usecase.GetPromiseHistoryListUseCase
 import com.yomo.yakgwa.util.UiState
-import com.skydoves.sandwich.onSuccess
-import com.skydoves.sandwich.retrofit.serialization.onErrorDeserialize
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
